aether/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
mem0ai>=0.1.0
httpx>=0.27.0,<1.0.0
pyngrok>=7.2.0,<8.0.0
